Refactored modules and bot interface
