+++ pkg/vk/content/content.go
@@ -1,19 +1,9 @@
+// Package provides interfaces for application
+// content services
 package content
 
 import (
 	"github.com/SevereCloud/vksdk/v2/object"
-	"github.com/sirupsen/logrus"
-)
-
-var packageLogFields = logrus.Fields{
-	"package": "content",
-}
-
-type AttachmentsType string
-
-const (
-	AudioType AttachmentsType = "audio"
-	PhotoType AttachmentsType = "photo"
 )
 
 type AttachmentsContentCollector interface {
